rename: 공간 api 타입 파일 정리
