+++ src/api/constants/categories.ts
@@ -1,3 +1,3 @@
-import { Category } from '@/pages/Main/types/spaces-type';
+import { Category } from '../types/activities';
 
 export const CATEGORIES: Category[] = ['문화 · 예술', '관광', '스포츠', '식음료', '웰빙', '투어', ''];
